+++ Soggins/docs/release-notes.md
@@ -5,6 +5,7 @@
 ## Upcoming release
 
 * Reduced Mysterious Ammo card's cost from 1 to 0.
+* Reduced the amount of Missile Malfunction given by the Repeated Mistakes artifact from 3 to 2.
 
 ## 1.5.0
 Released 22 January 2025.